Switch rendering and model logic to decimal coordinates
Migrated all geometry and rendering code from int to decimal types for coordinates and sizes. This enables support for floating-point values throughout the model, parsing, and SVG output. Updated model properties, parsing logic, code generation, renderable classes, and SVG helpers to use decimal. Adjusted sample data and point parsing to allow sub-pixel precision, improving rendering fidelity and flexibility.

+++ src/Steffi.Models/Builder/ModelBuilder.SetParentProperty.cs
@@ -11,12 +11,12 @@ public static partial class ModelBuilder
 		{
 			if (propertyName.Equals(nameof(CanvasParentProperties.X), StringComparison.InvariantCultureIgnoreCase))
 			{
-				canvasContainerProperties.X = int.Parse(value);
+				canvasContainerProperties.X = decimal.Parse(value);
 				return true;
 			}
 			else if (propertyName.Equals(nameof(CanvasParentProperties.Y), StringComparison.InvariantCultureIgnoreCase))
 			{
-				canvasContainerProperties.Y = int.Parse(value);
+				canvasContainerProperties.Y = decimal.Parse(value);
 				return true;
 			}
 		}
