+++ src/Steffi.Models/Containers/Properties/CanvasParentProperties.cs
@@ -2,6 +2,6 @@ namespace Steffi.Models.Containers.Properties;
 
 public class CanvasParentProperties : ParentProperties
 {
-	public int X { get; set; }
-	public int Y { get; set; }
+	public decimal X { get; set; }
+	public decimal Y { get; set; }
 }
