+++ src/Steffi.Renderers/Svg/Renderables/LineRenderable.cs
@@ -2,9 +2,9 @@
 
 namespace Steffi.Renderers.Svg.Renderables;
 
-internal class LineRenderable(int x1, int y1, int x2, int y2) : Renderable
+internal class LineRenderable(decimal x1, decimal y1, decimal x2, decimal y2) : Renderable
 {
-	public override (XElement Element, int Width, int Height) Render()
+	public override (XElement Element, decimal Width, decimal Height) Render()
 	{
 		return (SvgBuilder.Line(x1, y1, x2, y2), 0, 0);
 	}
